Started Events import

+++ src/main/kotlin/com/footiestats/statsengine/dtos/statsbomb/StatsBombLineupPlayer.kt
@@ -6,6 +6,6 @@ class StatsBombLineupPlayer(
     @JsonAlias("player_id") var playerId: Long,
     @JsonAlias("player_name") var playerName: String,
     @JsonAlias("player_nickname") var playerNickname: String?,
-    @JsonAlias("jersey_number") var jerseyNumber: Int,
+    @JsonAlias("jersey_number") var jerseyNumber: Byte,
     @JsonAlias("country") var country: StatsBombCountry?
 )
